refactor: avoid generic function with callback
