+++ src/common/campaign/post.ts
@@ -3,7 +3,7 @@ import { AuthContext } from '../../Context';
 import {
   ArticlePost,
   CampaignPost,
-  CampaignType,
+  CampaignState,
   Post,
   PostType,
   type ConnectionManager,
@@ -20,21 +20,30 @@ import { In } from 'typeorm';
 import { mapCloudinaryUrl } from '../cloudinary';
 import { pickImageUrl } from '../post';
 import { NotFoundError } from '../../errors';
-import { debeziumTimeToDate, updateFlagsStatement } from '../utils';
+import { debeziumTimeToDate, systemUser, updateFlagsStatement } from '../utils';
 import { getDiscussionLink } from '../links';
 import { skadiApiClient } from '../../integrations/skadi/api/clients';
 import { largeNumberFormat } from '../devcard';
 import { formatMailDate, addNotificationEmailUtm } from '../mailing';
 import { truncatePostToTweet } from '../twitter';
 import type { TemplateDataFunc } from '../../workers/newNotificationV2Mail';
-import { usdToCores } from '../number';
+import { coresToUsd, usdToCores } from '../number';
 
 import {
-  createNewCampaign,
-  startCampaign,
+  startCampaignTransferCores,
   validateCampaignArgs,
   type StartCampaignMutationArgs,
+  type StopCampaignProps,
 } from './common';
+import { randomUUID } from 'crypto';
+import {
+  UserTransaction,
+  UserTransactionProcessor,
+  UserTransactionStatus,
+  UserTransactionType,
+} from '../../entity/user/UserTransaction';
+import { addDays } from 'date-fns';
+import { capitalize } from 'lodash';
 
 export interface GQLPromotedPost
   extends Omit<
@@ -280,25 +289,114 @@ export const startCampaignPost = async (props: StartCampaignMutationArgs) => {
   checkPostAlreadyBoosted(post);
 
   const request = await ctx.con.transaction(async (manager) => {
-    const campaign = createNewCampaign(props, CampaignPost, {
-      postId,
-      type: CampaignType.Post,
+    const id = randomUUID();
+    const { budget, duration } = args;
+    const total = budget * duration;
+    const userId = ctx.userId;
+    const endedAt = addDays(new Date(), duration);
+
+    const campaign = await manager.getRepository(CampaignPost).save(
+      manager.getRepository(CampaignPost).create({
+        id,
+        flags: {
+          budget: total,
+          spend: 0,
+          users: 0,
+          clicks: 0,
+          impressions: 0,
+        },
+        userId,
+        referenceId: postId,
+        state: CampaignState.Active,
+        endedAt,
+        postId,
+      }),
+    );
+
+    const campaignId = campaign.id;
+
+    await skadiApiClient.startCampaign({
+      value: campaign.id,
+      type: campaign.type,
+      durationInDays: duration,
+      budget: coresToUsd(budget),
+      userId: campaign.userId,
     });
 
-    return startCampaign({
-      campaign,
-      manager,
-      args,
+    const userTransaction = await manager.getRepository(UserTransaction).save(
+      manager.getRepository(UserTransaction).create({
+        id: randomUUID(),
+        processor: UserTransactionProcessor.Njord,
+        receiverId: systemUser.id,
+        status: UserTransactionStatus.Success,
+        productId: null,
+        senderId: userId,
+        value: total,
+        valueIncFees: 0,
+        fee: 0,
+        request: ctx.requestMeta,
+        flags: { note: `Post Boost started` },
+        referenceId: campaignId,
+        referenceType: UserTransactionType.PostBoost,
+      }),
+    );
+
+    return startCampaignTransferCores({
       ctx,
-      onCampaignSaved: async () =>
-        manager
-          .getRepository(Post)
-          .update(
-            { id: postId },
-            { flags: updateFlagsStatement<Post>({ campaignId: campaign.id }) },
-          ),
+      manager,
+      campaignId,
+      userTransaction,
     });
   });
 
   return request.transaction;
+};
+
+export const stopCampaignPost = async ({
+  campaign,
+  ctx,
+}: StopCampaignProps) => {
+  const { id: campaignId, userId, referenceId } = campaign;
+
+  const { currentBudget } = await skadiApiClient.cancelCampaign({
+    campaignId,
+    userId,
+  });
+
+  const result = await ctx.con.transaction(async (manager) => {
+    const toRefund = parseFloat(currentBudget);
+
+    await manager
+      .getRepository(Post)
+      .update(
+        { id: referenceId },
+        { flags: updateFlagsStatement<Post>({ campaignId: null }) },
+      );
+
+    const userTransaction = await manager.getRepository(UserTransaction).save(
+      manager.getRepository(UserTransaction).create({
+        id: randomUUID(),
+        processor: UserTransactionProcessor.Njord,
+        receiverId: userId,
+        status: UserTransactionStatus.Success,
+        productId: null,
+        senderId: systemUser.id,
+        value: usdToCores(toRefund),
+        valueIncFees: 0,
+        fee: 0,
+        flags: { note: `${capitalize(campaign.type)} Boost refund` },
+        referenceId: campaignId,
+        referenceType: UserTransactionType.PostBoost,
+      }),
+    );
+
+    return startCampaignTransferCores({
+      ctx,
+      manager,
+      campaignId,
+      userTransaction,
+    });
+  });
+
+  return result.transaction;
 };
